requirements.txt
aiohttp==3.9.3
nextcord==2.6.0
nextcord_ext_ipc==2.2.1
Pillow==10.3.0
psutil==5.9.8
pymongo==4.6.2
python_Levenshtein==0.25.1
Requests==2.32.3
valaw==0.1.2
